work on teacher model

+++ segmentation/train.py
@@ -277,8 +277,7 @@ def train_one_epoch(model, train_loader, device, optimizer, loss_function_train,
         pred_scales = model(*input_data)
         # loss computation
         # losses = loss_function_train(pred_scales, torch.nn.functional.one_hot(target_scales, num_classes=19))
-        losses = loss_function_train(pred_scales, torch.nn.functional.one_hot(target_scales.long(),
-                                            num_classes=19).permute(0, 3, 1, 2))
+        losses = loss_function_train(pred_scales, torch.nn.functional.one_hot(target_scales, num_classes=19).permute(0, 3, 1, 2))
 
         loss_segmentation = sum(losses)
         total_loss = loss_segmentation
